third_party/WebKit/LayoutTests/fast/dom/shadow/shadow-element-inactive-expected.txt

CONSOLE WARNING: line 14: <shadow> doesn't work for APPLET element host.
CONSOLE WARNING: line 14: <shadow> doesn't work for DETAILS element host.
CONSOLE WARNING: line 14: <shadow> doesn't work for EMBED element host.
CONSOLE WARNING: line 14: <shadow> doesn't work for INPUT element host.
CONSOLE WARNING: line 14: <shadow> doesn't work for OBJECT element host.
CONSOLE WARNING: line 14: <shadow> doesn't work for SUMMARY element host.
We should not expose UA shadow nodes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


applet
PASS root2.firstChild.getDistributedNodes().length is 0
details
PASS root2.firstChild.getDistributedNodes().length is 0
embed
PASS root2.firstChild.getDistributedNodes().length is 0
input
PASS root2.firstChild.getDistributedNodes().length is 0
object
PASS root2.firstChild.getDistributedNodes().length is 0
summary
PASS root2.firstChild.getDistributedNodes().length is 0
PASS successfullyParsed is true

TEST COMPLETE
